+++ CHANGES.md
@@ -4,6 +4,8 @@ Changes
 1.3.5.4 (unreleased)
 --------------------
 
+- Add the "official" NamingSystem definitions.
+
 - The resource reference from the `MessageHeader` to the so called focal
   resource of the bundle needs to be versioned.